+++ 05-testing/02-angular-2025/solution/application/src/app/highlight.directive.ts
@@ -1,0 +1,19 @@
+import { Directive, ElementRef, Input, OnChanges } from '@angular/core';
+
+@Directive({
+  selector: '[appHighlight]',
+})
+export class HighlightDirective implements OnChanges {
+  defaultColor = 'rgb(211, 211, 211)';
+
+  @Input('appHighlight') bgColor = '';
+
+  constructor(private el: ElementRef) {
+    el.nativeElement.style.customProperty = true;
+  }
+
+  ngOnChanges(): void {
+    this.el.nativeElement.style.backgroundColor =
+      this.bgColor || this.defaultColor;
+  }
+}
